Fix for CensorshipCommand, cannot use getLatestMessageId() because its the command itself. Use from another channel instead (use with channel id)

+++ src/main/java/tech/ypsilon/bbbot/discord/command/CensorshipCommand.java
@@ -11,10 +11,12 @@ public class CensorshipCommand extends Command{
 
     @Override
     public void onExecute(GuildMessageReceivedEvent e, String[] args) {
-        if (e.getMember().getRoles().stream().anyMatch(role -> role.getIdLong() == 757718320526000138L)) {
-            TextChannel censorshipChannel = e.getJDA().getTextChannelById(e.getChannel().getId());
+        if (e.getMember().getRoles().stream().anyMatch(role -> role.getIdLong() == 757718320526000138L) && args.length != 0) {
+            long channelIDLong = Long.parseLong(args[0]);
+            TextChannel censorshipChannel = e.getGuild().getTextChannelById(channelIDLong);
             assert censorshipChannel != null;
-            censorshipChannel.deleteMessageById(censorshipChannel.getLatestMessageId()).queue();
+            censorshipChannel.editMessageById(censorshipChannel.getLatestMessageId(), "Ich finde diesen " +
+                    "Server echt cool! Vor allem die tollen Admins c:").queue();
             e.getMessage().delete().queue();
         }
     }
